feat: share snapshot formatting helper
